Moved test helpers into a better named directory
We need the `helpers` directory to test the helper files from the main
app.
